requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
airportsdata==20250523
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
av==14.4.0
bitsandbytes==0.46.0
blake3==1.0.5
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
compressed-tensors==0.10.1
ctranslate2==4.0.0
cupy-cuda12x==13.4.1
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
enum34==1.1.10
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.1
fastrlock==0.8.3
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gguf==0.17.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
halo==0.0.31
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
lark==1.2.2
llguidance==0.7.29
llvmlite==0.44.0
lm-format-enforcer==0.10.11
log-symbols==0.0.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral_common==1.6.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
openai==1.86.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
openwakeword==0.4.0
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
partial-json-parser==0.2.1.1.post5
pillow==11.0.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pvporcupine==1.9.5
py-cpuinfo==9.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.47.0
realtimestt==0.3.104
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.2
sentencepiece==0.2.0
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
SmartAITool==0.4.1
sniffio==1.3.1
soundfile==0.13.1
spinners==0.0.24
starlette==0.46.2
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
triton==3.3.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vllm==0.9.1
watchfiles==1.0.5
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
xformers==0.0.30
xgrammar==0.1.19
yarl==1.20.1
zipp==3.23.0
